requirements.txt
beautifulsoup4==4.13.4
requests==2.32.3
elasticsearch==8.9.0
neo4j==5.13.0
PyPDF2==3.0.1
pdfminer.six==20250506
dagster==1.4.13
dagster-webserver==1.4.13
dagster-postgres==0.20.13
pytest==8.4.0
pytest-asyncio==1.0.0
Faker==20.1.0
PyJWT==2.10.1
loguru==0.7.3
aiohttp==3.12.9
fastapi==0.115.12
httpx==0.28.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
pendulum==2.1.2
slack-sdk==3.26.1
pandas==2.3.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
elasticsearch-dsl==8.18.0
neo4j-driver==5.28.1
tika==3.1.0
alembic==1.16.1
pytest-mock==3.12.0
pytest-cov==4.1.0

# Standard Migration Tools
# Neo4j Migrations (Java-based, install separately)
# Download from: https://github.com/michael-simons/neo4j-migrations/releases

# Elasticsearch Evolution (Java-based, install separately) 
# Download from: https://github.com/senacor/elasticsearch-evolution/releases

# Python HTTP client for Elasticsearch migrations
requests>=2.31.0 